requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1
Werkzeug==3.1.3
email-validator==2.2.0
numpy==2.0.0
pymysql==1.1.1
